+++ server/src/engine/storage/v2/mod.rs
@@ -26,8 +26,11 @@
 
 use {
     self::{
-        impls::mdl_journal::{BatchStats, FullModel},
-        raw::journal::{JournalSettings, RepairResult},
+        impls::{
+            gns_log::{self, GNSEventLog},
+            mdl_journal::{BatchStats, FullModel, ModelDataAdapter},
+        },
+        raw::journal::{BatchAdapter, EventLogAdapter, JournalSettings, RepairResult},
     },
     super::{common::interface::fs::FileSystem, v1, SELoaded},
     crate::{
@@ -40,16 +43,12 @@ use {
             fractal::{context, FractalGNSDriver},
             storage::{
                 common::paths_v1,
-                v2::raw::journal::{self, JournalRepairMode},
-            },
-            txn::{
-                gns::{
-                    model::CreateModelTxn,
-                    space::CreateSpaceTxn,
-                    sysctl::{AlterUserTxn, CreateUserTxn},
+                v2::{
+                    impls::{gns_log::GNSAdapter, mdl_journal::ModelAdapter},
+                    raw::journal::{self, JournalRepairMode},
                 },
-                SpaceIDRef,
             },
+            txn::gns::sysctl::{AlterUserTxn, CreateUserTxn},
             RuntimeResult,
         },
         util,
@@ -63,18 +62,15 @@ pub(super) mod raw;
 pub const GNS_PATH: &str = v1::GNS_PATH;
 pub const DATA_DIR: &str = v1::DATA_DIR;
 
+/*
+    upgrade
+*/
+
 pub fn recreate(gns: GNSData) -> RuntimeResult<SELoaded> {
     context::set_dmsg("creating gns");
     let mut gns_driver = impls::gns_log::GNSDriver::create_gns()?;
-    // create all spaces
-    context::set_dmsg("creating all spaces");
-    for (space_name, space) in gns.idx().read().iter() {
-        FileSystem::create_dir_all(&paths_v1::space_dir(space_name, space.get_uuid()))?;
-        gns_driver.commit_event(CreateSpaceTxn::new(space.props(), &space_name, space))?;
-    }
-    // create all models
-    context::set_dmsg("creating all models");
-    for (model_id, model) in gns.idx_models().read().iter() {
+    gns_log::reinit_full::<true>(&mut gns_driver, &gns, |model_id, model| {
+        // re-initialize model
         let model_data = model.data();
         let space_uuid = gns.idx().read().get(model_id.space()).unwrap().get_uuid();
         FileSystem::create_dir_all(&paths_v1::model_dir(
@@ -89,23 +85,18 @@ pub fn recreate(gns: GNSData) -> RuntimeResult<SELoaded> {
             model_id.entity(),
             model_data.get_uuid(),
         ))?;
-        gns_driver.commit_event(CreateModelTxn::new(
-            SpaceIDRef::with_uuid(model_id.space(), space_uuid),
-            model_id.entity(),
-            model_data,
-        ))?;
         model_driver.commit_with_ctx(FullModel::new(model_data), BatchStats::new())?;
         model.driver().initialize_model_driver(model_driver);
-    }
-    // create all users
-    context::set_dmsg("creating all users");
-    for (user_name, user) in gns.sys_db().users().read().iter() {
-        gns_driver.commit_event(CreateUserTxn::new(&user_name, user.hash()))?;
-    }
+        Ok(())
+    })?;
     Ok(SELoaded {
         gns: GlobalNS::new(gns, FractalGNSDriver::new(gns_driver)),
     })
 }
+
+/*
+    initialize
+*/
 
 pub fn initialize_new(config: &Configuration) -> RuntimeResult<SELoaded> {
     FileSystem::create_dir_all(DATA_DIR)?;
@@ -126,10 +117,20 @@ pub fn initialize_new(config: &Configuration) -> RuntimeResult<SELoaded> {
     })
 }
 
+/*
+    restore
+*/
+
 pub fn restore(cfg: &Configuration) -> RuntimeResult<SELoaded> {
     let gns = GNSData::empty();
     context::set_dmsg("loading gns");
-    let mut gns_driver = impls::gns_log::GNSDriver::open_gns(&gns, JournalSettings::default())?;
+    let (mut gns_driver, gns_driver_stats) =
+        impls::gns_log::GNSDriver::open_gns(&gns, JournalSettings::default())?;
+    if gns_driver_stats.recommended_action().needs_compaction() {
+        gns_driver = journal::compact_journal::<true, EventLogAdapter<GNSEventLog>>(
+            GNS_PATH, gns_driver, &gns,
+        )?;
+    }
     let mut initialize_drivers = || {
         for (id, model) in gns.idx_models().write().iter_mut() {
             let model_data = model.data();
@@ -137,11 +138,18 @@ pub fn restore(cfg: &Configuration) -> RuntimeResult<SELoaded> {
             let model_data_file_path =
                 paths_v1::model_path(id.space(), space_uuid, id.entity(), model_data.get_uuid());
             context::set_dmsg(format!("loading model driver in {model_data_file_path}"));
-            let model_driver = impls::mdl_journal::ModelDriver::open_model_driver(
+            let (mut model_driver, mdl_stats) = impls::mdl_journal::ModelDriver::open_model_driver(
                 model_data,
                 &model_data_file_path,
                 JournalSettings::default(),
             )?;
+            if mdl_stats.recommended_action().needs_compaction() {
+                model_driver = journal::compact_journal::<true, BatchAdapter<ModelDataAdapter>>(
+                    &model_data_file_path,
+                    model_driver,
+                    model.data(),
+                )?;
+            }
             model.driver().initialize_model_driver(model_driver);
             unsafe {
                 // UNSAFE(@ohsayan): all pieces of data are upgraded by now, so vacuum
@@ -191,6 +199,10 @@ pub fn restore(cfg: &Configuration) -> RuntimeResult<SELoaded> {
         }
     }
 }
+
+/*
+    invoke repair
+*/
 
 pub fn repair() -> RuntimeResult<()> {
     // back up all files
@@ -256,4 +268,43 @@ fn print_repair_info(result: RepairResult, id: &str) {
             }
         }
     }
+}
+
+pub fn compact() -> RuntimeResult<()> {
+    let gns = GNSData::empty();
+    context::set_dmsg("reading GNS");
+    let stats = journal::read_journal::<GNSAdapter>(GNS_PATH, &gns, JournalSettings::default())?;
+    if !stats.recommended_action().needs_compaction() {
+        warn!("GNS does not need compaction");
+    }
+    journal::compact_journal_direct::<true, GNSAdapter, _>(GNS_PATH, None, &gns, true, |_| Ok(()))?;
+    for (id, model) in gns.idx_models().write().iter_mut() {
+        let model_data = model.data();
+        let space_uuid = gns.idx().read().get(id.space()).unwrap().get_uuid();
+        let model_data_file_path =
+            paths_v1::model_path(id.space(), space_uuid, id.entity(), model_data.get_uuid());
+        context::set_dmsg(format!("loading model driver in {model_data_file_path}"));
+        if !journal::read_journal::<ModelAdapter>(
+            &model_data_file_path,
+            model.data(),
+            JournalSettings::default(),
+        )?
+        .recommended_action()
+        .needs_compaction()
+        {
+            warn!(
+                "model {}.{} does not need compaction",
+                id.space(),
+                id.entity()
+            );
+        }
+        journal::compact_journal_direct::<true, ModelAdapter, _>(
+            &model_data_file_path,
+            None,
+            model.data(),
+            true,
+            |_| Ok(()),
+        )?;
+    }
+    Ok(())
 }